[loader] Move to an OO Loader, driving inference for image-classifier
This allows for the Loader to be used for inference on different kinds of data, e.g. inferring on text with a seq2seq model.
