+++ docs/AOT.md
@@ -17,10 +17,11 @@ bundles are object files that can be linked with some executable. On other
 architectures, the bundle may look completely different.
 
 This document demonstrates how to produce a bundle for the host CPU using the
-'loader' tool.  We use the flag `-emit-bundle` to specify the output directory.
+'image-classifier' tool.  We use the flag `-emit-bundle` to specify the output
+directory.
 
 ```
-loader image_file -image_mode=0to1 -m network_model_name -cpu -emit-bundle output_directory_name
+image-classifier image_file -image_mode=0to1 -m network_model_name -cpu -emit-bundle output_directory_name
 ```
 
 The command above would compile the neural network model from the
@@ -98,7 +99,7 @@ weights.
 This section describes the use of the CPU bundle. Other targets may have
 different interfaces.
 
-To integrate the artifacts generated by the loader into your project, you
+To integrate the artifacts generated by the image-classifier into your project, you
 generally need to do the following:
 * You need to link with the generated object file `network_model_name.o`.
 * You need to allocate the memory for constant weights variables,
@@ -115,9 +116,9 @@ variables area.
 
 ## A step-by-step example of the Resnet50 network model
 
-There are concrete examples of integrating a network model with a project.  You
-can find it in the `examples/compile_resnet50` directory in the Glow
-repository. Makefile with appropriate targets is provided for your convenience.
+There are concrete examples of integrating a network model with a project. You
+can find them in the `examples/bundles/` directory in the Glow repository.
+Makefiles with appropriate targets are provided for your convenience.
 
 ### Floating point network
 To build and run the example, you just need to execute:
@@ -128,9 +129,9 @@ your setup, primarily LOADER and GLOW_SRC vars.
 
 The makefile provides the following targets:
 * `download_weights`: it downloads the Resnet50 network model in the Caffe2 format.
-* `build/resnet50.o`: it generates the bundle files using the Glow loader as described above.
+* `build/resnet50.o`: it generates the bundle files using the Glow image-classifier as described above.
   The concrete command line looks like this:
-  `loader tests/images/imagenet/cat_285.png -image_mode=0to1 -m resnet50 -cpu -emit-bundle build`
+  `image-classifier tests/images/imagenet/cat_285.png -image_mode=0to1 -m resnet50 -cpu -emit-bundle build`
   It reads the network model from `resnet50` and generates the `resnet50.o`
   and `resnet50.weights` files into the `build` directory.
 * `build/main.o`:  it compiles the `resnet50_standalone.cpp` file, which is the main file of the project.
@@ -153,9 +154,9 @@ To build and run the example, you just need to execute:
 
 This run performs almost the same steps as non-quantized Resnet50 version
 except it emits bundle based on the quantization profile:
-`loader tests/images/imagenet/cat_285.png -image_mode=0to1 -m resnet50
+`image-classifier tests/images/imagenet/cat_285.png -image_mode=0to1 -m resnet50
 -load_profile=profile.yml -cpu -emit-bundle build`
 
-The `profile.yml` itself is captured at a prior step by executing loader with the `dump_profile` option:
-`loader tests/images/imagenet/*.png -image_mode=0to1 -m resnet50 -dump_profile=profile.yml`.
+The `profile.yml` itself is captured at a prior step by executing image-classifier with the `dump_profile` option:
+`image-classifier tests/images/imagenet/*.png -image_mode=0to1 -m resnet50 -dump_profile=profile.yml`.
 See the makefile for details.
