+++ docs/Quantization.md
@@ -64,14 +64,14 @@ from Resnet50.
 
 ### How to perform NN conversion
 
-Glow loader provides options to execute both profiling and conversion of a NN graph.
+The Glow loader tool provides options to execute both profiling and conversion of a NN graph.
 
 ```dump_profile=profile.yaml``` option is used to dump per node's output profile data
 into the ```profile.yaml``` file.
 This information can be used in the process of quantized conversion.
 For example, you can run the following command to capture profile for Resnet50.
 ```
-./bin/loader tests/images/*.png -image_mode=0to1 -m=resnet50 -dump_profile="profile.yaml"
+./bin/image-classifier tests/images/*.png -image_mode=0to1 -m=resnet50 -dump_profile="profile.yaml"
 ```
 
 ```load_profile=profile.yaml``` option is used to quantize graph based on the
@@ -81,7 +81,7 @@ the graph.
 For example, you can run the following command to load the profile and quantize
 the graph.
 ```
-./bin/loader tests/images/*.png -image_mode=0to1 -m=resnet50 -load_profile="profile.yaml"
+./bin/image-classifier tests/images/*.png -image_mode=0to1 -m=resnet50 -load_profile="profile.yaml"
 ```
 
 ## Compiler Optimizations
