+++ tools/loader/Loader.cpp
@@ -1,0 +1,231 @@
+/**
+ * Copyright (c) 2017-present, Facebook, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+#include "Loader.h"
+
+#include "glow/Base/Tensor.h"
+#include "glow/ExecutionEngine/ExecutionEngine.h"
+#include "glow/IR/IR.h"
+#include "glow/Quantization/Serialization.h"
+
+#include "llvm/Support/CommandLine.h"
+#include "llvm/Support/FileSystem.h"
+#include "llvm/Support/FormatVariadic.h"
+#include "llvm/Support/Timer.h"
+#include "llvm/Support/raw_ostream.h"
+
+using namespace glow;
+
+namespace {
+llvm::cl::OptionCategory loaderCat("Loader Options");
+
+llvm::cl::list<std::string> modelPathOpt(
+    "model",
+    llvm::cl::desc(
+        "Specify one of three:\n"
+        "1. Path to ONNX model file.\n"
+        "2. Two paths to Caffe2 model files: network structure and weight.\n"
+        "3. Path to directory with the Caffe2 network structure "
+        "<predict_net.pb> and weight <init_net.pb> files."),
+    llvm::cl::value_desc("modelPath"), llvm::cl::Required, llvm::cl::OneOrMore,
+    llvm::cl::cat(loaderCat));
+llvm::cl::alias modelPathAOpt("m", llvm::cl::desc("Alias for -model"),
+                              llvm::cl::aliasopt(modelPathOpt),
+                              llvm::cl::cat(loaderCat));
+
+llvm::cl::opt<bool>
+    verbose("verbose",
+            llvm::cl::desc("Specify whether to run with verbose output"),
+            llvm::cl::Optional, llvm::cl::cat(loaderCat));
+
+llvm::cl::opt<bool>
+    timeOpt("time",
+            llvm::cl::desc("Print timer output to stderr detailing how long it "
+                           "takes for the program to execute"),
+            llvm::cl::Optional, llvm::cl::cat(loaderCat));
+
+llvm::cl::opt<unsigned> iterationsOpt(
+    "iterations", llvm::cl::desc("Number of iterations to perform"),
+    llvm::cl::Optional, llvm::cl::init(1), llvm::cl::cat(loaderCat));
+
+llvm::cl::opt<std::string> dumpProfileFileOpt(
+    "dump_profile",
+    llvm::cl::desc("Perform quantization profiling for a given graph "
+                   "and dump result to the file."),
+    llvm::cl::value_desc("profile.yaml"), llvm::cl::Optional,
+    llvm::cl::cat(loaderCat));
+
+llvm::cl::opt<std::string> loadProfileFileOpt(
+    "load_profile",
+    llvm::cl::desc("Load quantization profile file and quantize the graph"),
+    llvm::cl::value_desc("profile.yaml"), llvm::cl::Optional,
+    llvm::cl::cat(loaderCat));
+
+llvm::cl::opt<BackendKind> ExecutionBackend(
+    llvm::cl::desc("Backend to use:"),
+    llvm::cl::values(clEnumValN(BackendKind::Interpreter, "interpreter",
+                                "Use interpreter"),
+                     clEnumValN(BackendKind::CPU, "cpu", "Use CPU"),
+                     clEnumValN(BackendKind::OpenCL, "opencl", "Use OpenCL")),
+    llvm::cl::init(BackendKind::Interpreter), llvm::cl::cat(loaderCat));
+
+/// Debugging options.
+llvm::cl::OptionCategory
+    modelExportCat("How to export the Glow Intermediate Representation/Graphs",
+                   "These options are for debugging the "
+                   "graphs by writing the IR/Graphs to "
+                   "given files/stdout");
+
+llvm::cl::opt<std::string> dumpGraphDAGFileOpt(
+    "dumpGraphDAG",
+    llvm::cl::desc("Specify the file to export the Graph in DOT format"),
+    llvm::cl::value_desc("file.dot"), llvm::cl::cat(modelExportCat));
+
+llvm::cl::opt<bool> dumpGraphOpt("dumpGraph",
+                                 llvm::cl::desc("Prints Graph to stdout"),
+                                 llvm::cl::cat(modelExportCat));
+
+llvm::cl::opt<std::string> dumpIRDAGFileOpt(
+    "dumpIRDAG",
+    llvm::cl::desc("Specify the file to export the IR in DOT format"),
+    llvm::cl::value_desc("file.dot"), llvm::cl::cat(modelExportCat));
+
+llvm::cl::opt<bool> dumpIROpt("dumpIR", llvm::cl::desc("Prints IR to stdout"),
+                              llvm::cl::cat(modelExportCat));
+
+/// Emit a bundle into the specified output directory.
+llvm::cl::opt<std::string>
+    emitBundle("emit-bundle",
+               llvm::cl::desc("Output directory for the bundle serialization"),
+               llvm::cl::cat(loaderCat));
+} // namespace
+
+bool glow::emittingBundle() { return !emitBundle.empty(); }
+
+static bool commandLineIsInvalid() {
+  if (!dumpProfileFileOpt.empty() && !loadProfileFileOpt.empty()) {
+    llvm::errs() << "Loader: the -" << dumpProfileFileOpt.ArgStr << " and -"
+                 << loadProfileFileOpt.ArgStr
+                 << " options may not be specified together.\n";
+    return true;
+  }
+  return false;
+}
+
+void Loader::compile() {
+  // Handle the request to profile the graph in preperation for quantization.
+  if (!dumpProfileFileOpt.empty()) {
+    // Perform the high-level optimizations before instrumenting the graph. This
+    // optimization phase will remove stuff like repetitive transpose operations
+    // perform CSE, etc.
+    ::optimize(F_, glow::CompilationMode::Infer);
+
+    // Instrument the graph to capture profiles for nodes' outputs.
+    ::profileQuantization(F_);
+  }
+
+  // Load the quantization profile and transform the graph.
+  if (!loadProfileFileOpt.empty()) {
+    // The profiled graph was optimized before it was instrumentated. In this
+    // part of the code we repeat the same transformation in order to create
+    // the same graph structure.
+    ::optimize(F_, glow::CompilationMode::Infer);
+
+    auto quantizationInfos = deserializeFromYaml(loadProfileFileOpt);
+
+    // Quantize the graph based on the captured profile.
+    quantization::generateQuantizedGraph(EE_, F_, quantizationInfos);
+  }
+
+  if (emittingBundle()) {
+    // Emit IR for the graph, compile it and save as a bundle.
+    EE_.save(CompilationMode::Infer, F_, emitBundle);
+  } else {
+    // Emit IR for the graph and compile it.
+    EE_.compile(CompilationMode::Infer, F_);
+  }
+
+  if (dumpGraphOpt) {
+    F_->dump();
+  }
+  if (!dumpGraphDAGFileOpt.empty()) {
+    F_->dumpDAG(dumpGraphDAGFileOpt.c_str());
+  }
+  if (dumpIROpt) {
+    EE_.getIR().dump();
+  }
+  if (!dumpIRDAGFileOpt.empty()) {
+    EE_.getIR().dumpDAG(dumpIRDAGFileOpt.c_str());
+  }
+}
+
+void Loader::runInference(llvm::ArrayRef<Variable *> variables,
+                          llvm::ArrayRef<Tensor *> tensors) {
+  assert(!emittingBundle() &&
+         "No inference is performed in the bundle generation mode.");
+
+  llvm::Timer timer("Infer", "Infer");
+  if (timeOpt) {
+    timer.startTimer();
+  }
+  for (unsigned i = 0; i < iterationsOpt; i++) {
+    EE_.run(variables, tensors);
+  }
+  if (timeOpt) {
+    timer.stopTimer();
+    llvm::outs() << llvm::formatv("Wall time per iteration (s): {0:f4}\n",
+                                  timer.getTotalTime().getWallTime() /
+                                      iterationsOpt);
+  }
+
+  if (!dumpProfileFileOpt.empty()) {
+    std::vector<NodeQuantizationInfo> QI =
+        quantization::generateNodeQuantizationInfos(F_);
+    serializeToYaml(dumpProfileFileOpt, QI);
+  }
+}
+
+Loader::Loader(int argc, char **argv) {
+  llvm::cl::ParseCommandLineOptions(
+      argc, argv,
+      " The Glow compiler\n\n"
+      "Glow is a compiler for neural network accelerators.\n");
+
+  if (commandLineIsInvalid()) {
+    std::exit(1);
+  }
+
+  if (modelPathOpt.size() > 2) {
+    llvm::errs() << "-model flag should have either 1 or 2 paths assigned. "
+                    "Please see flag's description.\n";
+    std::exit(1);
+  }
+
+  if (modelPathOpt.size() == 1) {
+    if (llvm::sys::fs::is_directory(*modelPathOpt.begin())) {
+      caffe2NetDescFilename_ = modelPathOpt[0] + "/predict_net.pb";
+      caffe2NetWeightFilename_ = modelPathOpt[0] + "/init_net.pb";
+    } else {
+      onnxModelFilename_ = modelPathOpt[0];
+    }
+  } else {
+    caffe2NetDescFilename_ = modelPathOpt[0];
+    caffe2NetWeightFilename_ = modelPathOpt[1];
+  }
+
+  EE_.setBackend(ExecutionBackend);
+  F_ = EE_.getModule().createFunction(modelPathOpt[0]);
+}
